+++ controllers/instancegroup_controller.go
@@ -48,9 +48,9 @@ import (
 // InstanceGroupReconciler reconciles an InstanceGroup object
 type InstanceGroupReconciler struct {
 	client.Client
+	SpotRecommendationTime float64
 	Log                    logr.Logger
 	ControllerConfPath     string
-	ControllerTemplatePath string
 	MaxParallel            int
 	Auth                   *InstanceGroupAuthenticator
 }
@@ -129,6 +129,13 @@ func (r *InstanceGroupReconciler) Reconcile(req ctrl.Request) (ctrl.Result, erro
 			return ctrl.Result{}, nil
 		}
 		r.Log.Error(err, "reconcile failed")
+		return ctrl.Result{}, err
+	}
+
+	if err = instanceGroup.Spec.Validate(); err != nil {
+		r.Log.Error(err, "reconcile failed")
+		instanceGroup.SetState(v1alpha1.ReconcileErr)
+		r.Update(context.Background(), instanceGroup)
 		return ctrl.Result{}, err
 	}
 
@@ -251,11 +258,17 @@ func (r *InstanceGroupReconciler) spotEventReconciler(obj handler.MapObject) []c
 		return nil
 	}
 
-	reason, exists, err := unstructured.NestedString(unstructuredObj, "reason")
-	if !exists || err != nil {
+	if reason, ok, _ := unstructured.NestedString(unstructuredObj, "reason"); ok {
+		if reason != kubeprovider.SpotRecommendationReason {
+			return nil
+		}
+	} else {
 		return nil
 	}
-	if reason != kubeprovider.SpotRecommendationReason {
+
+	creationTime := obj.Meta.GetCreationTimestamp()
+	minutesSince := time.Since(creationTime.Time).Minutes()
+	if minutesSince > r.SpotRecommendationTime {
 		return nil
 	}
 
@@ -279,10 +292,6 @@ func (r *InstanceGroupReconciler) spotEventReconciler(obj handler.MapObject) []c
 	instanceGroup.Name = awsprovider.GetTagValueByKey(tags, provisioners.TagInstanceGroupName)
 	instanceGroup.Namespace = awsprovider.GetTagValueByKey(tags, provisioners.TagInstanceGroupNamespace)
 	if instanceGroup.Name == "" || instanceGroup.Namespace == "" {
-		r.Log.Error(err,
-			"failed to map v1.event to scalinggroup",
-			"event", obj.Meta.GetName(),
-		)
 		return nil
 	}